user service : customer create / balance update edits

+++ user-service/src/main/java/com/dp/userservice/model/user/UserFinanceDTO.java
@@ -1,10 +1,11 @@
 package com.dp.userservice.model.user;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.List;
 
 public record UserFinanceDTO(
-        double balance,
+        BigDecimal balance,
         boolean automaticDebtDeduction,
         List<UserTransactionDTO> transactions
 ) {
