+++ user-service/src/main/java/com/dp/userservice/model/user/UserTransactionDTO.java
@@ -1,11 +1,12 @@
 package com.dp.userservice.model.user;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 public record UserTransactionDTO (
         Long id,
         ZonedDateTime datetime,
-        double sum,
+        BigDecimal sum,
         String type,
         String method
 ) {
